LayoutTests/media/modern-media-controls/media-controller/media-controller-toggle-compact-mode-expected.txt

Testing we toggle compact mode as we move past the threshold.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Controls width = 600px
mediaControls.classList.contains("compact") = false

Controls width = 241px
mediaControls.classList.contains("compact") = true

Controls width = 242px
mediaControls.classList.contains("compact") = false

PASS successfullyParsed is true

TEST COMPLETE
